# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  businessName: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  businessName: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface Sale {
  id: number;
  userId: number;
  productName: string;
  quantity: number;
  buyPrice: number;
  sellPrice: number;
  profit: number;
  /** @nullable */
  notes?: string | null;
  images: string[];
  /** @nullable */
  inventoryItemId?: number | null;
  saleDate: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSaleBody {
  productName: string;
  quantity: number;
  buyPrice: number;
  sellPrice: number;
  /** @nullable */
  notes?: string | null;
  images?: string[];
  /** @nullable */
  inventoryItemId?: number | null;
  saleDate: string;
}

export interface UpdateSaleBody {
  productName?: string;
  quantity?: number;
  buyPrice?: number;
  sellPrice?: number;
  /** @nullable */
  notes?: string | null;
  images?: string[];
  /** @nullable */
  inventoryItemId?: number | null;
  saleDate?: string;
}

export interface Expense {
  id: number;
  userId: number;
  title: string;
  amount: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  notes?: string | null;
  expenseDate: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateExpenseBody {
  title: string;
  amount: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  notes?: string | null;
  expenseDate: string;
}

export interface UpdateExpenseBody {
  title?: string;
  amount?: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  notes?: string | null;
  expenseDate?: string;
}

export interface InventoryItem {
  id: number;
  userId: number;
  name: string;
  quantity: number;
  buyPrice: number;
  sellPrice: number;
  /** @nullable */
  description?: string | null;
  images: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateInventoryItemBody {
  name: string;
  quantity: number;
  buyPrice: number;
  sellPrice: number;
  /** @nullable */
  description?: string | null;
  images?: string[];
}

export interface UpdateInventoryItemBody {
  name?: string;
  quantity?: number;
  buyPrice?: number;
  sellPrice?: number;
  /** @nullable */
  description?: string | null;
  images?: string[];
}

export interface UserSettings {
  id: number;
  userId: number;
  businessName: string;
  theme: string;
  currency: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateSettingsBody {
  businessName?: string;
  theme?: string;
  currency?: string;
}

export interface DashboardStats {
  totalRevenue: number;
  totalProfit: number;
  totalExpenses: number;
  netProfit: number;
  totalSales: number;
  totalItems: number;
}

export interface ChartDataPoint {
  label: string;
  revenue: number;
  profit: number;
}

export interface TopProduct {
  name: string;
  totalQuantity: number;
  totalRevenue: number;
  totalProfit: number;
}

export interface MonthlyReport {
  month: number;
  year: number;
  totalRevenue: number;
  totalProfit: number;
  totalExpenses: number;
  netProfit: number;
  totalSales: number;
  chartData: ChartDataPoint[];
  topProducts: TopProduct[];
  sales: Sale[];
  expenses: Expense[];
}

export type GetSalesParams = {
  /**
   * @nullable
   */
  startDate?: string | null;
  /**
   * @nullable
   */
  endDate?: string | null;
  /**
   * @nullable
   */
  month?: number | null;
  /**
   * @nullable
   */
  year?: number | null;
};

export type GetExpensesParams = {
  /**
   * @nullable
   */
  startDate?: string | null;
  /**
   * @nullable
   */
  endDate?: string | null;
  /**
   * @nullable
   */
  month?: number | null;
  /**
   * @nullable
   */
  year?: number | null;
};

export type GetDashboardStatsParams = {
  period: GetDashboardStatsPeriod;
};

export type GetDashboardStatsPeriod =
  (typeof GetDashboardStatsPeriod)[keyof typeof GetDashboardStatsPeriod];

export const GetDashboardStatsPeriod = {
  today: "today",
  week: "week",
  month: "month",
  year: "year",
} as const;

export type GetDashboardChartParams = {
  period: GetDashboardChartPeriod;
};

export type GetDashboardChartPeriod =
  (typeof GetDashboardChartPeriod)[keyof typeof GetDashboardChartPeriod];

export const GetDashboardChartPeriod = {
  today: "today",
  week: "week",
  month: "month",
  year: "year",
} as const;

export type GetMonthlyReportParams = {
  month: number;
  year: number;
};