# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Business Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: Get all expenses
      parameters:
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [today, week, month, year]
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/chart:
    get:
      operationId: getDashboardChart
      tags: [dashboard]
      summary: Get revenue and profit chart data
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [today, week, month, year]
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        businessName:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - businessName
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required:
        - id
        - businessName
        - email
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        amount:
          type: number
        category:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        expenseDate:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - title
        - amount
        - expenseDate
        - createdAt
        - updatedAt
    CreateExpenseBody:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        category:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        expenseDate:
          type: string
      required:
        - title
        - amount
        - expenseDate
    UpdateExpenseBody:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        category:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        expenseDate:
          type: string
...
    DashboardStats:
      type: object
      properties:
        totalRevenue:
          type: number
        totalProfit:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        totalSales:
          type: integer
        totalItems:
          type: integer
      required:
        - totalRevenue
        - totalProfit
        - totalExpenses
        - netProfit
        - totalSales
        - totalItems
    ChartDataPoint:
      type: object
      properties:
        label:
          type: string
        revenue:
          type: number
        profit:
          type: number
      required:
        - label
        - revenue
        - profit
...