# lib/api-zod/src/generated/types/createExpenseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseBody {
  title: string;
  amount: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  notes?: string | null;
  expenseDate: string;
}